move more dialogs into FormEntryDialogs
